fix: Make RequireJS config paths relative to workspace file
If multi-root workspaces are used, paths in variables `requireModuleSupport.modulePath` and `requireModuleSupport.configFile` are expected to be relative to the workspace root, to the directory with the `.code-workspace` file.

BREAKING CHANGE: If you used multi-root workspaces, you will need to modify paths in variables `requireModuleSupport.modulePath` and `requireModuleSupport.configFile` to be relative to the directory with the `.code-workspace` file. Because the support for multi-root workspaces wasn't complete, you had to use paths relative to a workspace folder. If the workspace folders had different directory nesting level in relation to the workspace root, it was impossible to configure RequireJS properly. The easiest solution is placing the RequireJS config file next to the `.code-workspace` file and setting `requireModuleSupport.modulePath` to `"."` and `requireModuleSupport.configFile` to the RequireJS config file name.

Paths inside the RequireJS config file are always supposed to be relative to the workspace root, which has not changed.